Add currentTime parameter to drawBands function
The drawBands function has been updated to accept an additional argument called currentTime. This argument is used to determine whether a band is currently playing, with color-coding being added to visually denote this on the interface. The call to this function has also been updated to pass in the current time.